+++ util/rand.h
@@ -1,4 +1,4 @@
-/* (c) Copyright 2004, 2005 Stijn van Dongen
+/*   (C) Copyright 2004, 2005, 2006 Stijn van Dongen
  *
  * This file is part of tingea.  You can redistribute and/or modify tingea
  * under the terms of the GNU General Public License; either version 2 of the
@@ -16,8 +16,8 @@
  *   *not* be suitable for cryptographic needs
 */
 
-unsigned int mcxSeed
-(  unsigned int seedlet
+unsigned long mcxSeed
+(  unsigned long seedlet
 )  ;
